+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/BulkPrintService.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.reform.finrem.caseorchestration.service;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.client.DocumentClient;
@@ -13,14 +14,23 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.model.document.BulkPrintRequ
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 
-import static uk.gov.hmcts.reform.finrem.caseorchestration.service.BulkPrintDocumentTranslator.approvedOrderCollection;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.service.BulkPrintDocumentTranslator.uploadOrder;
+import static java.util.Optional.ofNullable;
+import static uk.gov.hmcts.reform.finrem.caseorchestration.service.CommonFunction.getFirstMapValue;
+import static uk.gov.hmcts.reform.finrem.caseorchestration.service.CommonFunction.getLastMapValue;
 
 @Service
 @Slf4j
 public class BulkPrintService extends AbstractDocumentService {
+
+    private static final String DOCUMENT_FILENAME = "document_filename";
+    private static final String DOCUMENT_URL = "document_binary_url";
+    private static final String VALUE = "value";
+
+    @Value("${feature.approved-consent-order-notification-letter}")
+    private boolean featureApprovedConsentOrderNotificationLetter;
 
     @Autowired
     public BulkPrintService(DocumentClient documentClient,
@@ -30,7 +40,6 @@ public class BulkPrintService extends AbstractDocumentService {
     }
 
     public UUID sendForBulkPrint(final CaseDocument coverSheet, final CaseDetails caseDetails) {
-
         List<BulkPrintDocument> bulkPrintDocuments = new ArrayList<>();
 
         bulkPrintDocuments.add(
@@ -55,5 +64,82 @@ public class BulkPrintService extends AbstractDocumentService {
                 .letterType("FINANCIAL_REMEDY_PACK")
                 .bulkPrintDocuments(bulkPrintDocuments)
                 .build());
+    }
+
+    List<BulkPrintDocument> uploadOrder(Map<String, Object> data) {
+        log.info("Extracting 'uploadOrder' from case data for bulk print.");
+        List<BulkPrintDocument> bulkPrintDocuments = new ArrayList<>();
+        List<Map> documentList = ofNullable(data.get("uploadOrder"))
+            .map(i -> (List<Map>) i)
+            .orElse(new ArrayList<>());
+        if (documentList.size() > 0) {
+            Map<String, Object> value = ((Map) getLastMapValue.apply(documentList).get("value"));
+            bulkPrintDocuments.addAll(convertBulkPrintDocument(value, "DocumentLink"));
+        }
+        return bulkPrintDocuments;
+    }
+
+    List<BulkPrintDocument> approvedOrderCollection(Map<String, Object> data) {
+        log.info("Extracting 'approvedOrderCollection' from case data for bulk print.");
+        List<BulkPrintDocument> bulkPrintDocuments = new ArrayList<>();
+        List<Map> documentList = ofNullable(data.get("approvedOrderCollection"))
+            .map(i -> (List<Map>) i)
+            .orElse(new ArrayList<>());
+
+        if (documentList.size() > 0) {
+            Map<String, Object> value = ((Map) getFirstMapValue.apply(documentList).get("value"));
+            bulkPrintDocuments.addAll(convertBulkPrintDocument(value, "orderLetter"));
+            bulkPrintDocuments.addAll(convertBulkPrintDocument(value, "consentOrder"));
+
+            if (featureApprovedConsentOrderNotificationLetter) {
+                bulkPrintDocuments.addAll(convertBulkPrintDocument(value, "consentOrderApprovedNotificationLetter"));
+            }
+
+            bulkPrintDocuments.addAll(convertBulkPrintDocument(value, "pensionDocuments",
+                "uploadedDocument"));
+        }
+
+        return bulkPrintDocuments;
+    }
+
+    private List<BulkPrintDocument> convertBulkPrintDocument(Map<String, Object> data, String documentName) {
+        log.info("Extracting '{}' document from case data for bulk print.", documentName);
+        List<BulkPrintDocument> bulkPrintDocuments = new ArrayList<>();
+
+        Object documentLinkObj = data.get(documentName);
+
+        if (documentLinkObj != null) {
+            Map<String, Object> documentLink = (Map) documentLinkObj;
+            bulkPrintDocuments.add(BulkPrintDocument.builder()
+                .binaryFileUrl(documentLink.get(DOCUMENT_URL).toString())
+                .build());
+            log.info("Sending {} ({}) for bulk print.", documentName, documentLink.get(DOCUMENT_FILENAME));
+        }
+        return bulkPrintDocuments;
+    }
+
+    private List<BulkPrintDocument> convertBulkPrintDocument(Map<String, Object> data, String collectionName, String documentName) {
+        log.info("Extracting '{}' collection from case data for bulk print.", collectionName);
+
+        List<BulkPrintDocument> bulkPrintDocuments = new ArrayList<>();
+
+        List<Map> documentList = ofNullable(data.get(collectionName))
+            .map(i -> (List<Map>) i)
+            .orElse(new ArrayList<>());
+
+        for (Map<String, Object> document : documentList) {
+            Map<String, Object> value = ((Map) document.get(VALUE));
+
+            Object documentLinkObj = value.get(documentName);
+
+            if (documentLinkObj != null) {
+                Map<String, Object> documentLink = (Map) documentLinkObj;
+                bulkPrintDocuments.add(BulkPrintDocument.builder()
+                    .binaryFileUrl(documentLink.get(DOCUMENT_URL).toString())
+                    .build());
+                log.info("Sending {} ({}) for bulk print.", collectionName, documentLink.get(DOCUMENT_FILENAME));
+            }
+        }
+        return bulkPrintDocuments;
     }
 }
