+++ src/main/resources/application.properties
@@ -113,3 +113,6 @@ idam.health.url=${idam.url}/health
 idam.s2s-auth.url=${AUTH_PROVIDER_SERVICE_CLIENT_BASEURL:http://localhost:4502}
 idam.s2s-auth.services-allowed-to-update=${BSP_SERVICES_ALLOWED_TO_UPDATE:bulk_scan_orchestrator}
 idam.s2s-auth.services-allowed-to-validate=${BSP_SERVICES_ALLOWED_TO_VALIDATE:bulk_scan_processor}
+
+# Features
+feature.approved-consent-order-notification-letter=${FEATURE_APPROVED_CONSENT_ORDER_NOTIFICATION_LETTER:false}
