+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/BulkPrintServiceTest.java
@@ -4,15 +4,20 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
-import org.mockito.Mock;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.test.context.ActiveProfiles;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
+import uk.gov.hmcts.reform.finrem.caseorchestration.BaseServiceTest;
 import uk.gov.hmcts.reform.finrem.caseorchestration.client.DocumentClient;
 import uk.gov.hmcts.reform.finrem.caseorchestration.config.DocumentConfiguration;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocument;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.document.BulkPrintDocument;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.document.BulkPrintRequest;
 
 import java.io.InputStream;
+import java.util.List;
 import java.util.UUID;
 
 import static org.hamcrest.Matchers.is;
@@ -20,14 +25,19 @@ import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class BulkPrintServiceTest {
+@ActiveProfiles("test-bulk-print-service")
+public class BulkPrintServiceTest extends BaseServiceTest {
 
-    @Mock
+    @Autowired
     private DocumentClient documentClientMock;
 
-    private ObjectMapper mapper = new ObjectMapper();
+    @Autowired
+    private BulkPrintService bulkPrintService;
 
-    private BulkPrintService service;
+    @Value("${feature.approved-consent-order-notification-letter}")
+    private boolean featureApprovedConsentOrderNotificationLetter;
+
+    private ObjectMapper mapper = new ObjectMapper();
 
     private UUID letterId;
 
@@ -40,32 +50,43 @@ public class BulkPrintServiceTest {
         DocumentConfiguration config = new DocumentConfiguration();
         config.setApprovedConsentOrderTemplate("test_template");
         config.setApprovedConsentOrderFileName("test_file");
-        documentClientMock = mock(DocumentClient.class);
-        service = new BulkPrintService(documentClientMock, config, mapper);
     }
 
     @Test
     public void shouldSendForBulkPrintForApproved() throws Exception {
-
         when(documentClientMock.bulkPrint(bulkPrintRequestArgumentCaptor.capture())).thenReturn(letterId);
 
-        UUID bulkPrintLetterId = service.sendForBulkPrint(
+        UUID bulkPrintLetterId = bulkPrintService.sendForBulkPrint(
             new CaseDocument(), caseDetails());
 
-        assertThat(letterId, is(bulkPrintLetterId));
-        assertThat(bulkPrintRequestArgumentCaptor.getValue().getBulkPrintDocuments().size(), is(6));
+        assertThat(bulkPrintLetterId, is(letterId));
+        assertThat(bulkPrintRequestArgumentCaptor.getValue().getBulkPrintDocuments().size(),
+            is(featureApprovedConsentOrderNotificationLetter ? 6 : 5));
     }
 
     @Test
     public void shouldSendForBulkPrintForNotApproved() throws Exception {
-
         when(documentClientMock.bulkPrint(bulkPrintRequestArgumentCaptor.capture())).thenReturn(letterId);
 
-        UUID bulkPrintLetterId = service.sendForBulkPrint(
+        UUID bulkPrintLetterId = bulkPrintService.sendForBulkPrint(
             new CaseDocument(), caseDetailsForNonApproved());
 
-        assertThat(letterId, is(bulkPrintLetterId));
+        assertThat(bulkPrintLetterId, is(letterId));
         assertThat(bulkPrintRequestArgumentCaptor.getValue().getBulkPrintDocuments().size(), is(2));
+    }
+
+    @Test
+    public void shouldConvertDocument() throws Exception {
+        List<BulkPrintDocument> bulkPrintDocuments = bulkPrintService.uploadOrder(caseDetails().getData());
+
+        assertThat(bulkPrintDocuments.size(), is(1));
+    }
+
+    @Test
+    public void shouldConvertCollectionDocument() throws Exception {
+        List<BulkPrintDocument> bulkPrintDocuments = bulkPrintService.approvedOrderCollection(caseDetails().getData());
+
+        assertThat(bulkPrintDocuments.size(), is(featureApprovedConsentOrderNotificationLetter ? 5 : 4));
     }
 
     private CaseDetails caseDetails() throws Exception {
