+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/BulkPrintServiceTestConfiguration.java
@@ -1,0 +1,19 @@
+package uk.gov.hmcts.reform.finrem.caseorchestration.service;
+
+import org.mockito.Mockito;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.context.annotation.Profile;
+import uk.gov.hmcts.reform.finrem.caseorchestration.client.DocumentClient;
+
+@Profile("test-bulk-print-service")
+@Configuration
+public class BulkPrintServiceTestConfiguration {
+
+    @Bean
+    @Primary
+    public DocumentClient documentClient() {
+        return Mockito.mock(DocumentClient.class);
+    }
+}
