+++ src/test/resources/application.properties
@@ -64,3 +64,6 @@ idam.health.url=${idam.url}/health
 idam.s2s-auth.url=false
 idam.s2s-auth.services-allowed-to-update=${BSP_SERVICES_ALLOWED_TO_UPDATE:bulk_scan_orchestrator}
 idam.s2s-auth.services-allowed-to-validate=${BSP_SERVICES_ALLOWED_TO_VALIDATE:bulk_scan_processor}
+
+# Features
+feature.approved-consent-order-notification-letter=${FEATURE_APPROVED_CONSENT_ORDER_NOTIFICATION_LETTER:false}
